+++ common/network/Encrypter.cpp
@@ -66,7 +66,7 @@ std::string Encrypter::GetPublicKey()
     ByteQueue queue;
     public_key_.Save(queue);
 
-    size_t length = queue.CurrentSize();
+    size_t length = static_cast<size_t>(queue.CurrentSize());
     std::unique_ptr<char[]> outbuf(new char [length]);
     queue.Get((byte*)outbuf.get(), length);
 
@@ -86,7 +86,7 @@ std::string Encrypter::GetPrivateKey()
     ByteQueue queue;
     private_key_.Save(queue);
 
-    size_t length = queue.CurrentSize();
+    size_t length = static_cast<size_t>(queue.CurrentSize());
     std::unique_ptr<char[]> outbuf(new char [length]);
     queue.Get((byte*)outbuf.get(), length);
